Work in progress

+++ src/ticks.cpp
@@ -1,3 +1,3 @@
 #include "ticks.hpp"
 
-uint32_t ticks() noexcept { return SDL_GetTicks(); }
+uint64_t ticks() noexcept { return SDL_GetTicks(); }
